backend/requirements.txt
# Core Dependencies
fastapi==0.115.0
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
uvicorn[standard]==0.31.0
python-multipart==0.0.12
tenacity==8.5.0
starlette==0.38.0

# OpenAI-compatible client surface
openai==1.51.0

# Vertex AI authentication helpers
google-generativeai==0.8.3
google-auth>=2.41.1

# Async Support
httpx==0.27.2
aiofiles==24.1.0
sse-starlette==2.1.3

# Database
aiosqlite==0.17.0

# Vector Search (SQLite extension for hybrid retrieval)
sqlite-vec>=0.1.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Telegram Integration
telethon==1.41.2

# Reddit Integration
asyncpraw>=7.8.1

# JSON Processing
json-repair==0.25.2

# Environment & Config
python-dotenv==1.0.1

# Development Dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
ruff==0.6.9

# Optional: For better development experience
ipython==8.28.0
rich==13.8.1
apscheduler==3.10.4
